structure around RestyResolver
closes #5
